cloudrun/bybit_trading/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
pybit>=5.8
httpx>=0.27
websockets>=12.0
